# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Project Pilot API",
    "version": "0.1.2"
  },
  "paths": {
...
    "/api/settings/github": {
      "get": {
        "tags": [
          "settings"
        ],
        "summary": "Get Github Settings",
        "operationId": "get_github_settings_api_settings_github_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GithubSettingsRead"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "settings"
        ],
        "summary": "Put Github Settings",
        "operationId": "put_github_settings_api_settings_github_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GithubSettingsUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GithubSettingsRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/settings/github/test": {
      "post": {
        "tags": [
          "settings"
        ],
        "summary": "Post Github Test",
        "operationId": "post_github_test_api_settings_github_test_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/GithubTestRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GithubTestResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/settings/translation": {
      "get": {
        "tags": [
          "settings"
        ],
        "summary": "Get Translation Settings",
        "operationId": "get_translation_settings_api_settings_translation_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TranslationSettingsRead"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "settings"
        ],
        "summary": "Put Translation Settings",
        "operationId": "put_translation_settings_api_settings_translation_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TranslationSettingsUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TranslationSettingsRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/settings/ai/config": {
      "get": {
        "tags": [
          "settings"
        ],
        "summary": "Get Ai Config",
        "operationId": "get_ai_config_api_settings_ai_config_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AiConfigRead"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "settings"
        ],
        "summary": "Put Ai Config",
        "operationId": "put_ai_config_api_settings_ai_config_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AiConfigUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AiConfigRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/settings/ai": {
      "get": {
        "tags": [
          "settings"
        ],
        "summary": "Get Ai Settings",
        "operationId": "get_ai_settings_api_settings_ai_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AiSettingsRead"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "settings"
        ],
        "summary": "Put Ai Settings",
        "operationId": "put_ai_settings_api_settings_ai_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AiSettingsUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AiSettingsRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/settings/ai/test": {
      "post": {
        "tags": [
          "settings"
        ],
        "summary": "Post Ai Test",
        "operationId": "post_ai_test_api_settings_ai_test_post",
        "parameters": [
          {
            "name": "provider_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Provider Id"
            }
          },
          {
            "name": "scenario_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Scenario Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AiTestResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AiConfigRead": {
        "properties": {
          "providers": {
            "items": {
              "$ref": "#/components/schemas/AiProviderRead"
            },
            "type": "array",
            "title": "Providers"
          },
          "default_provider_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Default Provider Id"
          },
          "scenarios": {
            "additionalProperties": {
              "$ref": "#/components/schemas/AiScenarioBinding"
            },
            "type": "object",
            "title": "Scenarios"
          },
          "scenario_labels": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Scenario Labels"
          },
          "supported_providers": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Supported Providers"
          }
        },
        "type": "object",
        "title": "AiConfigRead"
      },
      "AiConfigUpdate": {
        "properties": {
          "providers": {
            "items": {
              "$ref": "#/components/schemas/AiProviderWrite"
            },
            "type": "array",
            "title": "Providers"
          },
          "default_provider_id": {
            "type": "string",
            "title": "Default Provider Id"
          },
          "scenarios": {
            "additionalProperties": {
              "$ref": "#/components/schemas/AiScenarioBinding"
            },
            "type": "object",
            "title": "Scenarios"
          }
        },
        "type": "object",
        "required": [
          "providers",
          "default_provider_id"
        ],
        "title": "AiConfigUpdate"
      },
      "AiProviderRead": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "preset_id": {
            "type": "string",
            "title": "Preset Id"
          },
          "provider": {
            "type": "string",
            "title": "Provider"
          },
          "base_url": {
            "type": "string",
            "title": "Base Url"
          },
          "models": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Models"
          },
          "default_model": {
            "type": "string",
            "title": "Default Model",
            "default": ""
          },
          "has_api_key": {
            "type": "boolean",
            "title": "Has Api Key",
            "default": false
          },
          "api_key_preview": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Api Key Preview"
          },
          "api_key_length": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Api Key Length"
          },
          "is_default": {
            "type": "boolean",
            "title": "Is Default",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "preset_id",
          "provider",
          "base_url"
        ],
        "title": "AiProviderRead"
      },
      "AiProviderWrite": {
        "properties": {
          "id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id",
            "description": "省略则新建"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "preset_id": {
            "type": "string",
            "title": "Preset Id"
          },
          "provider": {
            "type": "string",
            "title": "Provider",
            "default": "openai_compatible"
          },
          "base_url": {
            "type": "string",
            "title": "Base Url"
          },
          "models": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Models"
          },
          "default_model": {
            "type": "string",
            "title": "Default Model",
            "default": ""
          },
          "api_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Api Key",
            "description": "写入数据库；省略表示不修改；传 null 或空字符串表示清除"
          }
        },
        "type": "object",
        "required": [
          "name",
          "preset_id",
          "base_url"
        ],
        "title": "AiProviderWrite"
      },
      "AiScenarioBinding": {
        "properties": {
          "provider_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider Id",
            "description": "绑定的供应商 id"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model",
            "description": "该场景使用的模型"
          }
        },
        "type": "object",
        "title": "AiScenarioBinding"
      },
      "AiSettingsRead": {
        "properties": {
          "provider": {
            "type": "string",
            "title": "Provider",
            "description": "LLM Provider"
          },
          "preset_id": {
            "type": "string",
            "title": "Preset Id",
            "description": "平台预设 id（如 minimax-cn）"
          },
          "base_url": {
            "type": "string",
            "title": "Base Url",
            "description": "OpenAI 兼容 API Base URL"
          },
          "model": {
            "type": "string",
            "title": "Model",
            "description": "模型名称"
          },
          "has_api_key": {
            "type": "boolean",
            "title": "Has Api Key",
            "description": "是否存在可用 API Key（数据库或环境变量）"
          },
          "api_key_preview": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Api Key Preview",
            "description": "仅当 Key 保存在数据库时返回末 4 位"
          },
          "api_key_length": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Api Key Length",
            "description": "仅当 Key 保存在数据库时返回字符长度"
          },
          "supported_providers": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Supported Providers",
            "description": "可选 Provider 列表"
          },
          "default_provider": {
            "type": "string",
            "title": "Default Provider",
            "default": "openai_compatible"
          },
          "default_preset_id": {
            "type": "string",
            "title": "Default Preset Id",
            "default": "minimax-cn"
          },
          "default_base_url": {
            "type": "string",
            "title": "Default Base Url",
            "default": "https://api.minimaxi.com/v1"
          },
          "default_model": {
            "type": "string",
            "title": "Default Model",
            "default": "MiniMax-M2.5-highspeed"
          }
        },
        "type": "object",
        "required": [
          "provider",
          "preset_id",
          "base_url",
          "model",
          "has_api_key"
        ],
        "title": "AiSettingsRead"
      },
      "AiSettingsUpdate": {
        "properties": {
          "provider": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider",
            "description": "LLM Provider"
          },
          "preset_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preset Id",
            "description": "平台预设 id"
          },
          "base_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Base Url",
            "description": "Base URL"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model",
            "description": "模型名称"
          },
          "api_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Api Key",
            "description": "写入数据库；省略表示不修改；传 null 或空字符串表示清除"
          }
        },
        "type": "object",
        "title": "AiSettingsUpdate"
      },
      "AiTestResponse": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "sample": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sample"
          }
        },
        "type": "object",
        "required": [
          "ok"
        ],
        "title": "AiTestResponse"
      },
...
      "GithubSettingsRead": {
        "properties": {
          "has_token": {
            "type": "boolean",
            "title": "Has Token",
            "description": "是否存在可用 Token（数据库或环境变量）"
          },
          "token_preview": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Token Preview",
            "description": "仅当 Token 保存在数据库时返回末 4 位；仅从环境变量读取时不返回末位"
          },
          "token_length": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Token Length",
            "description": "仅当 Token 保存在数据库时返回字符长度；仅从环境变量读取时不返回"
          }
        },
        "type": "object",
        "required": [
          "has_token"
        ],
        "title": "GithubSettingsRead",
        "description": "GET 响应：不返回完整 Token。"
      },
      "GithubSettingsUpdate": {
        "properties": {
          "token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Token",
            "description": "写入数据库；传 null 或空字符串表示清除数据库中的 Token"
          }
        },
        "type": "object",
        "title": "GithubSettingsUpdate"
      },
      "GithubTestRequest": {
        "properties": {
          "token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Token",
            "description": "可选：测试指定 Token；省略则测试当前生效 Token（数据库或环境变量）"
          }
        },
        "type": "object",
        "title": "GithubTestRequest"
      },
      "GithubTestResponse": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "ok"
        ],
        "title": "GithubTestResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "TranslationSettingsRead": {
        "properties": {
          "provider": {
            "type": "string",
            "title": "Provider",
            "description": "翻译 Provider（首版固定 google）"
          },
          "target_lang": {
            "type": "string",
            "title": "Target Lang",
            "description": "目标语言 BCP-47"
          },
          "supported_target_langs": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Supported Target Langs",
            "description": "可选目标语言列表"
          }
        },
        "type": "object",
        "required": [
          "provider",
          "target_lang"
        ],
        "title": "TranslationSettingsRead"
      },
      "TranslationSettingsUpdate": {
        "properties": {
          "target_lang": {
            "type": "string",
            "maxLength": 16,
            "minLength": 2,
            "title": "Target Lang"
          }
        },
        "type": "object",
        "required": [
          "target_lang"
        ],
        "title": "TranslationSettingsUpdate"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}